Added in elastic energy calculation and cleaned up module
