+++ pymatgen/analysis/elasticity/tests/test_elastic.py
@@ -6,7 +6,7 @@ import os
 
 import numpy as np
 from pymatgen.analysis.elasticity.elastic import ElasticTensor
-from pymatgen.analysis.elasticity.strain import Strain, IndependentStrain
+from pymatgen.analysis.elasticity.strain import Strain, IndependentStrain, Deformation
 from pymatgen.analysis.elasticity.stress import Stress
 from pymatgen.util.testing import PymatgenTest
 import warnings
@@ -132,5 +132,23 @@ class ElasticTensorTest(PymatgenTest):
             et_from_sd = ElasticTensor.from_stress_dict(stress_dict)
         self.assertArrayAlmostEqual(et_from_sd.symmetrized.round(2),
                                     self.elastic_tensor_1)
+
+    def test_energy(self):
+
+        film_elac = ElasticTensor([
+            [324.32,  187.3,   170.92,    0.,      0.,      0.],
+            [187.3,   324.32,  170.92,    0.,      0.,      0.],
+            [170.92,  170.92,  408.41,    0.,      0.,      0.],
+            [0.,      0.,      0.,    150.73,    0.,      0.],
+            [0.,      0.,      0.,      0.,    150.73,    0.],
+            [0.,      0.,      0.,      0.,      0.,    238.74]])
+
+        dfm = Deformation([[ -9.86004855e-01,2.27539582e-01,-4.64426035e-17],
+            [ -2.47802121e-01,-9.91208483e-01,-7.58675185e-17],
+            [ -6.12323400e-17,-6.12323400e-17,1.00000000e+00]])
+
+        self.assertAlmostEqual(film_elac.energy_density(dfm.green_lagrange_strain),
+            0.000125664672793)
+
 if __name__ == '__main__':
     unittest.main()
